[Feature #159236210] Change the way user adds a meal (#23)
Initially, when the user tried to make a nutrition plan, it was a bit uncomfortable to add a time for the meal and then have to add the food to that meal. It was a bit cumbersome to use two forms.

This commit enables users to set everything at once (time, ingredient and amount).